New stacks that use stages
