+++ lib/data-stack.ts
@@ -1,85 +1,72 @@
-import { CfnOutput, RemovalPolicy, Stack, StackProps } from 'aws-cdk-lib';
-import {
-  Table,
-  BillingMode,
-  AttributeType,
-  StreamViewType,
-  ProjectionType
-} from 'aws-cdk-lib/aws-dynamodb';
-import { Bucket, BlockPublicAccess } from 'aws-cdk-lib/aws-s3';
+import * as cdk from 'aws-cdk-lib';
+import * as ddb from 'aws-cdk-lib/aws-dynamodb';
+import * as s3 from 'aws-cdk-lib/aws-s3';
+
 import { Construct } from 'constructs';
 
-export interface StreamDataProps extends StackProps {
-  readonly environment: string;
+export interface DataStackProps extends cdk.StackProps {
+  environmentName: string;
 }
 
-export class DataStack extends Stack {
-  constructor(scope: Construct, id: string, props: StreamDataProps) {
+export class DataStack extends cdk.Stack {
+  constructor(scope: Construct, id: string, props: DataStackProps) {
     super(scope, id, props);
 
-    const streamDataTable = new Table(this, `StreamData-${props.environment}`, {
-      tableName: `Stream-Data-${props.environment}`,
-      billingMode: BillingMode.PAY_PER_REQUEST,
-      partitionKey: {
-        name: 'pk',
-        type: AttributeType.STRING
-      },
-      pointInTimeRecovery: true,
-      sortKey: {
-        name: 'sk',
-        type: AttributeType.STRING
-      },
-      stream: StreamViewType.NEW_IMAGE
-    });
+    const streamDataTable = new ddb.Table(
+      this,
+      `StreamData-${props.environmentName}`,
+      {
+        tableName: `Stream-Data-${props.environmentName}`,
+        billingMode: ddb.BillingMode.PAY_PER_REQUEST,
+        partitionKey: {
+          name: 'pk',
+          type: ddb.AttributeType.STRING
+        },
+        pointInTimeRecovery: true,
+        sortKey: {
+          name: 'sk',
+          type: ddb.AttributeType.STRING
+        },
+        stream: ddb.StreamViewType.NEW_IMAGE
+      }
+    );
 
     streamDataTable.addGlobalSecondaryIndex({
       indexName: 'gsi1',
       partitionKey: {
         name: 'gsi_pk1',
-        type: AttributeType.STRING
+        type: ddb.AttributeType.STRING
       },
       sortKey: {
         name: 'gsi_sk1',
-        type: AttributeType.STRING
+        type: ddb.AttributeType.STRING
       },
-      projectionType: ProjectionType.ALL
+      projectionType: ddb.ProjectionType.ALL
     });
 
     streamDataTable.addGlobalSecondaryIndex({
       indexName: 'gsi2',
       partitionKey: {
         name: 'gsi_pk2',
-        type: AttributeType.STRING
+        type: ddb.AttributeType.STRING
       },
       sortKey: {
         name: 'gsi_sk2',
-        type: AttributeType.STRING
+        type: ddb.AttributeType.STRING
       },
-      projectionType: ProjectionType.ALL
+      projectionType: ddb.ProjectionType.ALL
     });
 
-    new CfnOutput(this, `TableArnExport-${props.environment}`, {
-      value: streamDataTable.tableArn,
-      description: 'The name of the table',
-      exportName: `kb-db-table-arn-${props.environment}`
-    });
-
-    const dataBucket = new Bucket(
+    const dataBucket = new s3.Bucket(
       this,
-      `StreamDataBucket-${props.environment}`,
+      `StreamDataBucket-${props.environmentName}`,
       {
-        bucketName: `stream-data-${props.environment}`,
+        bucketName: `stream-data-${props.environmentName}`,
         publicReadAccess: false,
-        blockPublicAccess: BlockPublicAccess.BLOCK_ALL,
+        blockPublicAccess: s3.BlockPublicAccess.BLOCK_ALL,
         autoDeleteObjects: true,
-        removalPolicy: RemovalPolicy.DESTROY
+        removalPolicy: cdk.RemovalPolicy.DESTROY
       }
     );
-
-    new CfnOutput(this, `StreamDataBucketNameExport-${props.environment}`, {
-      value: dataBucket.bucketName,
-      description: 'The name of the stream data bucket',
-      exportName: `kb-stream-data-bucket-name-${props.environment}`
-    });
   }
 }
