+++ package.json
@@ -15,7 +15,7 @@
     "@types/aws-lambda": "^8.10.140",
     "@types/jest": "^29.5.11",
     "@types/node": "20.10.8",
-    "aws-cdk": "2.121.1",
+    "aws-cdk": "2.167.1",
     "aws-sdk-client-mock": "^4.0.1",
     "aws-sdk-mock": "^6.0.4",
     "esbuild": "^0.21.5",
@@ -32,7 +32,7 @@
     "@aws-sdk/client-ssm": "^3.600.0",
     "@aws-sdk/util-dynamodb": "^3.665.0",
     "@types/youtube": "^0.0.50",
-    "aws-cdk-lib": "2.161.1",
+    "aws-cdk-lib": "2.167.1",
     "constructs": "^10.0.0",
     "moment": "^2.30.1",
     "node-fetch": "^3.3.2",
